+++ statham/converters.py
@@ -1,5 +1,6 @@
 from typing import Type
 
+from statham.exceptions import ValidationError
 from statham.validators import NotPassed
 
 
@@ -30,5 +31,19 @@ def map_instantiate(model: Type):
         if isinstance(list_data, NotPassed):
             return list_data
         return [_factory(data) for data in list_data]
+
+    return _convert
+
+
+def any_of_instantiate(*models: Type):
+    def _convert(data):
+        if not isinstance(data, (dict, list)):
+            return data
+        for model in models:
+            try:
+                return instantiate(model)(data)
+            except (TypeError, ValidationError):
+                continue
+        raise ValidationError.from_composition(models, data)
 
     return _convert